server/requirements.txt
drf_yasg==1.20.0
user_agents==2.2.0
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==4.6.0
django-suit==2.0a1
django-select2==7.7.1
drf-writable-nested==0.6.3
django-filter==22.1
mysqlclient==2.1.1
pycryptodome==3.15.0
django-crontab==0.7.1
